refactor: move files around to fit new unicode detection
